recoveryiq/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
anthropic==0.18.0
langgraph==0.0.26
langchain-anthropic==0.1.1
pydantic==2.5.3
python-dotenv==1.0.0
opencv-python==4.9.0.80
scipy==1.12.0
numpy==1.26.4
requests==2.31.0
